Hello Visitor, this is a TODO List Website
This website was created using HTML, CSS, JavaScript in which we can add, freeze or remove a task in list. A Task is only can be removed if it is marked as completed. This project was made during completion of my course on Coursera. Link to Course: https://www.coursera.org/learn/duke-programming-web